Add files via upload
it is version 3 of the math game, it is the higher level math game by adding the Double digit plus double digit function, which make it more difficult. and in this version you should click the "enter" button to check the number you input, you can input double digit number.